feat(aquarium): Simplify publishing cycle
Before, adding a new entry would create a new issue for the given week
in case it didn't exist. Derived from this model, there was a need to
keep track of the status of each issue {published, ready, draft}.

Now, new entries don't add a new issue. This means that the new issue is
created on command, when publishing. This simplifies the process in two
ways: removes the need for a ready state and removes the need for a
predecided day of publication.

Signed-off-by: Arnau Siches <asiches@gmail.com>